refactor(schema): split aggregator.rs by responsibility (#1132)
Behavior-preserving module split of schema/aggregator.rs (~2,185 lines)
into a directory module per epic #1116:

- aggregator/mod.rs  - shared model (config, result, errors, ParsedSchema)
  + orchestration (discover/parse_file dispatch, two-pass apply_schemas)
  + pub use re-exports unchanged + #[cfg(test)] test_support fixtures
- aggregator/cql.rs  - CQL file handling (extract_use/create_keyspace,
  parse_cql_file) + CQL tests
- aggregator/json.rs - JSON schema handling (Json* structs, parse_json_file,
  parse_*_format, convert_* methods) + JSON tests

The existing 'pub mod aggregator;' in schema/mod.rs auto-resolves to the new
mod.rs; the public surface (SchemaAggregator, AggregatorConfig, LoadResult,
SchemaLoadError, LoadErrorType, SchemaLoadWarning) is unchanged. Tests follow
their code (count preserved: 27 active + 3 pre-existing #[ignore]); shared
fixtures live in test_support. No semantic changes.
